fix: userid超过Int上限导致崩溃，优化部分逻辑 opt：继续添加设置界面 version 0.3

+++ app/src/main/java/com/ghhccghk/musicplay/data/login/getLoginQr.kt
@@ -33,5 +33,5 @@ data class KeyData(
     val pic: String,
     val status: Int,
     val token: String,
-    val userid: Int
+    val userid: Long
 )
